build: explicit release builds
The build system now accecpts a new flag called "release" to indicate
that the build is release and not a snapshot. This will be important
for a new version scheme where feature branches will no longer build
release but snapshots and integrations build will also only produce
a snapshot but a dedicated nightly release job will create a release.